+++ src/main/java/com/gamemetricbackend/domain/user/dto/request/SignupRequestDto.java
@@ -1,4 +1,4 @@
-package com.gamemetricbackend.domain.user.dto;
+package com.gamemetricbackend.domain.user.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
